Reorganize panel usage and documented new configuration
