+++ web/client/configs/localConfig.json
@@ -812,9 +812,9 @@
           },
           { "name": "QueryPanel",
             "cfg": {
-              "useEmbeddedMap": true,
               "toolsOptions": {
-                "hideCrossLayer": true
+                "hideCrossLayer": true,
+                "useEmbeddedMap": true
               },
               "spatialPanelExpanded": false,
               "spatialOperations": [
