+++ web/client/plugins/QueryPanel.jsx
@@ -20,8 +20,7 @@ import { changeDrawingStatus } from '../actions/draw';
 import { getLayerCapabilities } from '../actions/layerCapabilities';
 import { queryPanelSelector } from '../selectors/controls';
 import { applyFilter, discardCurrentFilter, storeCurrentFilter } from '../actions/layerFilter';
-import MapComp from './queryPanelWithMap/MapComp';
-import SpatialFilterCustom from './queryPanelWithMap/SpatialFilter';
+import MapComp from './querypanel/MapComp';
 
 import {
     changeGroupProperties,
@@ -286,15 +285,6 @@ class QueryPanel extends React.Component {
     };
 
     renderQueryPanel = () => {
-        const MapComponent = this.props.mapComp;
-        if (this.props.useEmbeddedMap) {
-            standardItems.spatial = [{
-                id: "spatialFilter",
-                plugin: SpatialFilterCustom,
-                cfg: {},
-                position: 1
-            }];
-        }
         return (<div className="mapstore-query-builder">
             <SmartQueryForm
                 queryPanelEnabled={this.props.queryPanelEnabled}
@@ -311,11 +301,6 @@ class QueryPanel extends React.Component {
                 selectedLayer={this.props.selectedLayer}
                 standardItems={standardItems}
             />
-            {this.props.useEmbeddedMap ?
-                <div className="mapstore-query-map">
-                    <MapComponent/>
-                </div>
-                : null}
             <Portal>
                 <ResizableModal
                     fade
@@ -417,12 +402,13 @@ class QueryPanel extends React.Component {
  *   - srsName {string} The projection of the requested features fetched via wfs
  * Plugin acts as container and by default it have three panels: "AttributesFilter", "SpatialFilter" and "CrossLayerFilter" (see "standardItems" variable)
  * Panels can be customized by injection from another plugins (see example below).
- * Targets available for injection: "start", "attributes", "afterAttributes", "spatial", "afterSpatial", "layers", "end".
+ * Targets available for injection: "start", "attributes", "afterAttributes", "spatial", "afterSpatial", "layers", "end", "map"
 
  * @prop {object[]} cfg.spatialOperations: The list of geometric operations use to create the spatial filter.<br/>
  * @prop {boolean} cfg.toolsOptions.hideCrossLayer force cross layer filter panel to hide (when is not used or not usable)
  * @prop {boolean} cfg.toolsOptions.hideAttributeFilter force attribute filter panel to hide (when is not used or not usable). In general any `hide${CapitailizedItemId}` works to hide a particular panel of the query panel.
  * @prop {boolean} cfg.toolsOptions.hideSpatialFilter force spatial filter panel to hide (when is not used or not usable)
+ * @prop {boolean} cfg.toolsOptions.useEmbeddedMap if spatial filter panel is present, this option allows to use the embedded map instead of the map plugin
  *
  * @example
  * // This example configure a layer with polygons geometry as spatial filter method
