+++ web/client/plugins/querypanel/MapComp.jsx
