+++ web/client/plugins/querypanel/MapWithDraw.jsx
