+++ web/client/plugins/querypanel/SpatialFilter.jsx
@@ -5,8 +5,10 @@
  * This source code is licensed under the BSD-style license found in the
  * LICENSE file in the root directory of this source tree.
  */
+import React from 'react';
 import {connect} from "react-redux";
 import {mapSelector} from "../../selectors/map";
+import {getMapConfigSelector} from "../../selectors/queryform";
 import {bindActionCreators} from "redux";
 import {
     changeDwithinValue,
@@ -21,14 +23,12 @@ import {
 import {changeDrawingStatus} from "../../actions/draw";
 import SpatialFilterComponent from "../../components/data/query/SpatialFilter";
 
-const SpatialFilter = connect((state) => {
+const BaseSpatialFilter = connect((state) => {
     return {
         useMapProjection: state.queryform.useMapProjection,
         spatialField: state.queryform.spatialField,
         showDetailsPanel: state.queryform.showDetailsPanel,
-        spatialPanelExpanded: state.queryform.spatialPanelExpanded,
-        zoom: (mapSelector(state) || {}).zoom,
-        projection: (mapSelector(state) || {}).projection
+        spatialPanelExpanded: state.queryform.spatialPanelExpanded
     };
 }, dispatch => {
     return {
@@ -48,5 +48,25 @@ const SpatialFilter = connect((state) => {
         }, dispatch)
     };
 })(SpatialFilterComponent);
+
+/**
+ * Connected to the Map plugin
+ */
+const SpatialFilterMapPlugin = connect((state) => ({
+    zoom: (mapSelector(state) || {}).zoom,
+    projection: (mapSelector(state) || {}).projection
+}))(BaseSpatialFilter);
+
+/**
+ * Connected to the embedded map of the query panel
+ */
+export const SpatialFilterEmbeddedMap = connect((state) => ({
+    zoom: (getMapConfigSelector(state) || {}).zoom,
+    projection: (getMapConfigSelector(state) || {}).projection
+}))(BaseSpatialFilter);
+
+export const SpatialFilter = ({useEmbeddedMap, ...props}) => {
+    return useEmbeddedMap ? <SpatialFilterEmbeddedMap {...props}/> : <SpatialFilterMapPlugin {...props}/>;
+};
 
 export default SpatialFilter;
