+++ web/client/plugins/querypanel/enhancers/mapEnhancer.js
